refactor: update method add shoot

+++ src/main/java/com/teamobi/mobiarmy2/fight/IMapManager.java
@@ -26,5 +26,5 @@ public interface IMapManager {
 
     void addNewTiles(MapTile mapTile);
 
-    short[] getRandomPosition(short leftMargin, short rightMargin, short topMargin, short bottomMargin);
+    short[] getRandomPosition(int leftMargin, int rightMargin, int topMargin, int bottomMargin);
 }
